pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "miniflight"
version = "0.1.0"
description = "Physics engine"
dependencies = [
    "numpy>=1.24",
    "matplotlib",
    "hidapi",
    "gymnasium",
]

[tool.setuptools.packages.find]
include = ["miniflight", "common", "sim", "target"]